// packages/ops/src/gen/types/opsOverview.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Lemma Ops API
 * OpenAPI spec version: 0.1.0
 */
import type { OpsOverviewOutbox } from "./opsOverviewOutbox.js";
import type { OpsOverviewQueue } from "./opsOverviewQueue.js";

export interface OpsOverview {
  outbox: OpsOverviewOutbox;
  queue: OpsOverviewQueue;
}