// packages/ops/src/gen/types/opsQueueJob.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Lemma Ops API
 * OpenAPI spec version: 0.1.0
 */
import type { OpsQueueJobData } from "./opsQueueJobData.js";
import type { OpsQueueJobOutput } from "./opsQueueJobOutput.js";

export interface OpsQueueJob {
  /** @minLength 1 */
  id: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  state: string;
  /** @minimum 0 */
  retryCount: number;
  /** @minimum 0 */
  retryLimit: number;
  /** @nullable */
  data: OpsQueueJobData;
  /** @nullable */
  output: OpsQueueJobOutput;
  /** @nullable */
  createdOn: string | null;
  /** @nullable */
  startedOn: string | null;
  /** @nullable */
  completedOn: string | null;
}